Refactors, cleans and moves some resources to the new files.
