+++ pkg/occlient/occlient_test.go
@@ -21,14 +21,12 @@ import (
 	componentlabels "github.com/openshift/odo/pkg/component/labels"
 	"github.com/openshift/odo/pkg/config"
 	"github.com/openshift/odo/pkg/devfile/adapters/common"
-	"github.com/openshift/odo/pkg/log"
 	"github.com/openshift/odo/pkg/testingutil"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/apimachinery/pkg/watch"
 	ktesting "k8s.io/client-go/testing"
 )
@@ -200,29 +198,6 @@ func fakeBuildStatus(status buildv1.BuildPhase, buildName string) *buildv1.Build
 			Name:      buildName,
 		},
 		Status: buildv1.BuildStatus{
-			Phase: status,
-		},
-	}
-}
-
-func fakeEventStatus(podName string, eventWarningMessage string, count int32) *corev1.Event {
-	return &corev1.Event{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: podName,
-		},
-		Type:    "Warning",
-		Count:   count,
-		Reason:  eventWarningMessage,
-		Message: "Foobar",
-	}
-}
-
-func fakePodStatus(status corev1.PodPhase, podName string) *corev1.Pod {
-	return &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: podName,
-		},
-		Status: corev1.PodStatus{
 			Phase: status,
 		},
 	}
@@ -1172,63 +1147,6 @@ func TestGetExposedPorts(t *testing.T) {
 	}
 }
 
-func TestGetSecret(t *testing.T) {
-	tests := []struct {
-		name       string
-		secretNS   string
-		secretName string
-		wantErr    bool
-		want       *corev1.Secret
-	}{
-		{
-			name:       "Case: Valid request for retrieving a secret",
-			secretNS:   "",
-			secretName: "foo",
-			want: &corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "foo",
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name:       "Case: Invalid request for retrieving a secret",
-			secretNS:   "",
-			secretName: "foo2",
-			want: &corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "foo",
-				},
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fakeClient, fakeClientSet := FakeNew()
-
-			// Fake getting Secret
-			fakeClientSet.Kubernetes.PrependReactor("get", "secrets", func(action ktesting.Action) (bool, runtime.Object, error) {
-				if tt.want.Name != tt.secretName {
-					return true, nil, fmt.Errorf("'get' called with a different secret name")
-				}
-				return true, tt.want, nil
-			})
-
-			returnValue, err := fakeClient.GetSecret(tt.secretName, tt.secretNS)
-
-			// Check for validating return value
-			if err == nil && returnValue != tt.want {
-				t.Errorf("error in return value got: %v, expected %v", returnValue, tt.want)
-			}
-
-			if !tt.wantErr == (err != nil) {
-				t.Errorf("\nclient.GetSecret(secretNS, secretName) unexpected error %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestCreateServiceBinding(t *testing.T) {
 	tests := []struct {
 		name        string
@@ -1572,72 +1490,6 @@ func TestUnlinkSecret(t *testing.T) {
 	}
 }
 
-func TestListSecrets(t *testing.T) {
-
-	tests := []struct {
-		name       string
-		secretList corev1.SecretList
-		output     []corev1.Secret
-		wantErr    bool
-	}{
-		{
-			name: "Case 1: Ensure secrets are properly listed",
-			secretList: corev1.SecretList{
-				Items: []corev1.Secret{
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "secret1",
-						},
-					},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "secret2",
-						},
-					},
-				},
-			},
-			output: []corev1.Secret{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "secret1",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "secret2",
-					},
-				},
-			},
-
-			wantErr: false,
-		},
-	}
-
-	for _, tt := range tests {
-		client, fakeClientSet := FakeNew()
-
-		fakeClientSet.Kubernetes.PrependReactor("list", "secrets", func(action ktesting.Action) (bool, runtime.Object, error) {
-			return true, &tt.secretList, nil
-		})
-
-		secretsList, err := client.ListSecrets("")
-
-		if !reflect.DeepEqual(tt.output, secretsList) {
-			t.Errorf("expected output: %#v,got: %#v", tt.secretList, secretsList)
-		}
-
-		if err == nil && !tt.wantErr {
-			if len(fakeClientSet.Kubernetes.Actions()) != 1 {
-				t.Errorf("expected 1 action in ListSecrets got: %v", fakeClientSet.Kubernetes.Actions())
-			}
-		} else if err == nil && tt.wantErr {
-			t.Error("test failed, expected: false, got true")
-		} else if err != nil && !tt.wantErr {
-			t.Errorf("test failed, expected: no error, got error: %s", err.Error())
-		}
-	}
-}
-
 func TestGetImageStream(t *testing.T) {
 	tests := []struct {
 		name           string
@@ -1950,366 +1802,6 @@ func TestWaitForBuildToFinish(t *testing.T) {
 		})
 	}
 
-}
-
-func TestCollectEvents(t *testing.T) {
-	tests := []struct {
-		name                string
-		podName             string
-		eventWarningMessage string
-	}{
-		{
-			name:                "Case 1: Collect an arbitrary amount of events",
-			podName:             "ruby",
-			eventWarningMessage: "Fake event warning message",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			// Create a fake client
-			fakeClient, fakeClientSet := FakeNew()
-			fakeEventWatch := watch.NewRaceFreeFake()
-			podSelector := fmt.Sprintf("deploymentconfig=%s", tt.podName)
-
-			// Create a fake event status / watch reactor for faking the events we are collecting
-			fakeEvent := fakeEventStatus(tt.podName, tt.eventWarningMessage, 10)
-			go func(event *corev1.Event) {
-				fakeEventWatch.Add(event)
-			}(fakeEvent)
-
-			fakeClientSet.Kubernetes.PrependWatchReactor("events", func(action ktesting.Action) (handled bool, ret watch.Interface, err error) {
-				return true, fakeEventWatch, nil
-			})
-
-			// Create a test spinner / variables / quit variable for the go channel
-			spinner := log.Spinner("Test spinner")
-			events := make(map[string]corev1.Event)
-			quit := make(chan int)
-			go fakeClient.CollectEvents(podSelector, events, spinner, quit)
-
-			// Sleep in order to make sure we actually collect some events
-			time.Sleep(2 * time.Second)
-			close(quit)
-
-			// We make sure to lock in order to prevent race conditions when retrieving the events (since they are a pointer
-			// by default since we pass in a map)
-			mu.Lock()
-			if len(events) == 0 {
-				t.Errorf("Expected events, got none")
-			}
-			mu.Unlock()
-
-			// Collect the first event in the map
-			var firstEvent corev1.Event
-			for _, val := range events {
-				firstEvent = val
-			}
-
-			if !strings.Contains(firstEvent.Reason, tt.eventWarningMessage) {
-				t.Errorf("expected warning message: '%s' in event message: '%+v'", tt.eventWarningMessage, firstEvent.Reason)
-			}
-
-		})
-	}
-}
-
-// NOTE: We do *not* collection the amount of actions taken in this function as there could be any number of fake
-// 'event' actions that are happening in the background.
-func TestWaitAndGetPod(t *testing.T) {
-	tests := []struct {
-		name                string
-		podName             string
-		status              corev1.PodPhase
-		wantEventWarning    bool
-		wantErr             bool
-		eventWarningMessage string
-	}{
-		{
-			name:             "Case 1: Pod running",
-			podName:          "ruby",
-			status:           corev1.PodRunning,
-			wantEventWarning: false,
-			wantErr:          false,
-		},
-		{
-			name:             "Case 2: Pod failed",
-			podName:          "ruby",
-			status:           corev1.PodFailed,
-			wantEventWarning: false,
-			wantErr:          true,
-		},
-		{
-			name:             "Case 3: Pod unknown",
-			podName:          "ruby",
-			status:           corev1.PodUnknown,
-			wantEventWarning: false,
-			wantErr:          true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			fakeClient, fakeClientSet := FakeNew()
-			fakePodWatch := watch.NewRaceFreeFake()
-
-			// Watch for Pods
-			fakePod := fakePodStatus(tt.status, tt.podName)
-			go func(pod *corev1.Pod) {
-				fakePodWatch.Modify(pod)
-			}(fakePod)
-
-			// Prepend watch reactor (beginning of the chain)
-			fakeClientSet.Kubernetes.PrependWatchReactor("pods", func(action ktesting.Action) (handled bool, ret watch.Interface, err error) {
-				return true, fakePodWatch, nil
-			})
-
-			podSelector := fmt.Sprintf("deploymentconfig=%s", tt.podName)
-
-			pod, err := fakeClient.WaitAndGetPod(podSelector, corev1.PodRunning, "Waiting for component to start")
-
-			if !tt.wantErr == (err != nil) {
-				t.Errorf("client.WaitAndGetPod(string) unexpected error %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if err == nil {
-				if pod.Name != tt.podName {
-					t.Errorf("pod name is not matching to expected name, expected: %s, got %s", tt.podName, pod.Name)
-				}
-			}
-
-		})
-	}
-}
-
-func TestWaitAndGetSecret(t *testing.T) {
-
-	tests := []struct {
-		name       string
-		secretName string
-		namespace  string
-		wantErr    bool
-	}{
-		{
-			name:       "Case 1: no error expected",
-			secretName: "ruby",
-			namespace:  "dummy",
-			wantErr:    false,
-		},
-
-		{
-			name:       "Case 2: error expected",
-			secretName: "",
-			namespace:  "dummy",
-			wantErr:    true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			fkclient, fkclientset := FakeNew()
-			fkWatch := watch.NewFake()
-
-			// Change the status
-			go func() {
-				fkWatch.Modify(&corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: tt.secretName,
-					},
-				})
-			}()
-
-			fkclientset.Kubernetes.PrependWatchReactor("secrets", func(action ktesting.Action) (handled bool, ret watch.Interface, err error) {
-				if len(tt.secretName) == 0 {
-					return true, nil, fmt.Errorf("error watching secret")
-				}
-				return true, fkWatch, nil
-			})
-
-			pod, err := fkclient.WaitAndGetSecret(tt.secretName, tt.namespace)
-
-			if !tt.wantErr == (err != nil) {
-				t.Errorf(" client.WaitAndGetSecret(string, string) unexpected error %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(fkclientset.Kubernetes.Actions()) != 1 {
-				t.Errorf("expected 1 action in WaitAndGetSecret got: %v", fkclientset.Kubernetes.Actions())
-			}
-
-			if err == nil {
-				if pod.Name != tt.secretName {
-					t.Errorf("secret name is not matching to expected name, expected: %s, got %s", tt.secretName, pod.Name)
-				}
-			}
-		})
-	}
-}
-
-func TestCreateService(t *testing.T) {
-	tests := []struct {
-		name             string
-		commonObjectMeta metav1.ObjectMeta
-		containerPorts   []corev1.ContainerPort
-		wantErr          bool
-		existingDC       appsv1.DeploymentConfig
-	}{
-		{
-			name:       "Test case: with valid commonObjectName and containerPorts",
-			existingDC: *fakeDeploymentConfig("foo", "", nil, nil, t),
-			commonObjectMeta: metav1.ObjectMeta{
-				Name: "nodejs",
-				Labels: map[string]string{
-					"app":                        "apptmp",
-					"app.kubernetes.io/instance": "ruby",
-					"app.kubernetes.io/name":     "ruby",
-					"app.kubernetes.io/part-of":  "apptmp",
-				},
-				Annotations: map[string]string{
-					"app.openshift.io/vcs-uri":                "https://github.com/openshift/ruby",
-					"app.kubernetes.io/component-source-type": "git",
-				},
-			},
-			containerPorts: []corev1.ContainerPort{
-				{
-					Name:          "8080-tcp",
-					ContainerPort: 8080,
-					Protocol:      corev1.ProtocolTCP,
-				},
-				{
-					Name:          "9100-udp",
-					ContainerPort: 9100,
-					Protocol:      corev1.ProtocolUDP,
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fkclient, fkclientset := FakeNew()
-
-			ownerReference := metav1.OwnerReference{
-				APIVersion: "apps.openshift.io/v1",
-				Kind:       "DeploymentConfig",
-				Name:       tt.existingDC.Name,
-				UID:        tt.existingDC.UID,
-			}
-
-			_, err := fkclient.CreateService(tt.commonObjectMeta, tt.containerPorts, ownerReference)
-
-			tt.commonObjectMeta.SetOwnerReferences(append(tt.commonObjectMeta.GetOwnerReferences(), ownerReference))
-
-			if err == nil && !tt.wantErr {
-				if len(fkclientset.Kubernetes.Actions()) != 1 {
-					t.Errorf("expected 1 Kubernetes.Actions() in CreateService, got: %v", fkclientset.ImageClientset.Actions())
-				}
-				createdSvc := fkclientset.Kubernetes.Actions()[0].(ktesting.CreateAction).GetObject().(*corev1.Service)
-				if !reflect.DeepEqual(tt.commonObjectMeta, createdSvc.ObjectMeta) {
-					t.Errorf("ObjectMeta does not match the expected name, expected: %v, got: %v", tt.commonObjectMeta, createdSvc.ObjectMeta)
-				}
-				if !reflect.DeepEqual(tt.commonObjectMeta.Name, createdSvc.Spec.Selector["deploymentconfig"]) {
-					t.Errorf("selector value does not match the expected name, expected: %s, got: %s", tt.commonObjectMeta.Name, createdSvc.Spec.Selector["deploymentconfig"])
-				}
-				for _, port := range tt.containerPorts {
-					found := false
-					for _, servicePort := range createdSvc.Spec.Ports {
-						if servicePort.Port == port.ContainerPort {
-							found = true
-							if servicePort.Protocol != port.Protocol {
-								t.Errorf("service protocol does not match the expected name, expected: %s, got: %s", port.Protocol, servicePort.Protocol)
-							}
-							if servicePort.Name != port.Name {
-								t.Errorf("service name does not match the expected name, expected: %s, got: %s", port.Name, servicePort.Name)
-							}
-							if servicePort.TargetPort != intstr.FromInt(int(port.ContainerPort)) {
-								t.Errorf("target port does not match the expected name, expected: %v, got: %v", intstr.FromInt(int(port.ContainerPort)), servicePort.TargetPort)
-							}
-						}
-					}
-					if found == false {
-						t.Errorf("expected service port %s not found in the created Service", tt.name)
-						break
-					}
-				}
-			} else if err == nil && tt.wantErr {
-				t.Error("error was expected, but no error was returned")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, no error was expected, but got unexpected error: %s", err)
-			}
-		})
-	}
-}
-
-func TestGetDeploymentConfigsFromSelector(t *testing.T) {
-	tests := []struct {
-		name     string
-		selector string
-		label    map[string]string
-		wantErr  bool
-	}{
-		{
-			name:     "true case",
-			selector: "app.kubernetes.io/part-of=app",
-			label: map[string]string{
-				"app.kubernetes.io/part-of": "app",
-			},
-			wantErr: false,
-		},
-		{
-			name:     "true case",
-			selector: "app.kubernetes.io/part-of=app1",
-			label: map[string]string{
-				"app.kubernetes.io/part-of": "app",
-			},
-			wantErr: false,
-		},
-	}
-
-	listOfDC := appsv1.DeploymentConfigList{
-		Items: []appsv1.DeploymentConfig{
-			{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						"app.kubernetes.io/part-of": "app",
-					},
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fakeClient, fakeClientSet := FakeNew()
-
-			fakeClientSet.AppsClientset.PrependReactor("list", "deploymentconfigs", func(action ktesting.Action) (bool, runtime.Object, error) {
-				if !reflect.DeepEqual(action.(ktesting.ListAction).GetListRestrictions().Labels.String(), tt.selector) {
-					return true, nil, fmt.Errorf("labels not matching with expected values, expected:%s, got:%s", tt.selector, action.(ktesting.ListAction).GetListRestrictions())
-				}
-				return true, &listOfDC, nil
-			})
-			dc, err := fakeClient.GetDeploymentConfigsFromSelector(tt.selector)
-
-			if len(fakeClientSet.AppsClientset.Actions()) != 1 {
-				t.Errorf("expected 1 AppsClientset.Actions() in GetDeploymentConfigsFromSelector, got: %v", fakeClientSet.AppsClientset.Actions())
-			}
-
-			if tt.wantErr == false && err != nil {
-				t.Errorf("test failed, %#v", dc[0].Labels)
-			}
-
-			for _, dc1 := range dc {
-				if !reflect.DeepEqual(dc1.Labels, tt.label) {
-					t.Errorf("labels are not matching with expected labels, expected: %s, got %s", tt.label, dc1.Labels)
-				}
-			}
-
-		})
-	}
 }
 
 func TestCreateServiceInstance(t *testing.T) {
@@ -2668,106 +2160,6 @@ func TestGetClusterPlansFromServiceName(t *testing.T) {
 			} else if err != nil && !tt.wantErr {
 				t.Errorf("test failed, expected: no error, got error: %s", err.Error())
 			}
-		})
-	}
-}
-
-func TestGetDeploymentConfigLabelValues(t *testing.T) {
-	type args struct {
-		deploymentConfigList appsv1.DeploymentConfigList
-		expectedOutput       []string
-	}
-	tests := []struct {
-		applicationName string
-		name            string
-		args            args
-		wantErr         bool
-		actions         int
-	}{
-		{
-			name:            "Case 1 - Retrieve list",
-			applicationName: "app",
-			args: args{
-				expectedOutput: []string{"app", "app2"},
-				deploymentConfigList: appsv1.DeploymentConfigList{
-					Items: []appsv1.DeploymentConfig{
-						{
-							ObjectMeta: metav1.ObjectMeta{
-								Labels: map[string]string{
-									"app.kubernetes.io/part-of": "app",
-								},
-							},
-						},
-						{
-							ObjectMeta: metav1.ObjectMeta{
-								Labels: map[string]string{
-									"app.kubernetes.io/part-of": "app2",
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: false,
-			actions: 1,
-		},
-		{
-			name:            "Case 1 - Retrieve list, different order",
-			applicationName: "app",
-			args: args{
-				expectedOutput: []string{"app", "app2"},
-				deploymentConfigList: appsv1.DeploymentConfigList{
-					Items: []appsv1.DeploymentConfig{
-						{
-							ObjectMeta: metav1.ObjectMeta{
-								Labels: map[string]string{
-									"app.kubernetes.io/part-of": "app2",
-								},
-							},
-						},
-						{
-							ObjectMeta: metav1.ObjectMeta{
-								Labels: map[string]string{
-									"app.kubernetes.io/part-of": "app",
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: false,
-			actions: 1,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			fakeClient, fakeClientSet := FakeNew()
-
-			fakeClientSet.AppsClientset.PrependReactor("list", "deploymentconfigs", func(action ktesting.Action) (bool, runtime.Object, error) {
-				return true, &tt.args.deploymentConfigList, nil
-			})
-
-			// Run function GetServiceInstanceLabelValues
-			list, err := fakeClient.GetDeploymentConfigLabelValues(applabels.ApplicationLabel, applabels.ApplicationLabel)
-
-			if err == nil && !tt.wantErr {
-
-				// Compare arrays
-				if !reflect.DeepEqual(list, tt.args.expectedOutput) {
-					t.Errorf("expected %s output, got %s", tt.args.expectedOutput, list)
-				}
-
-				if (len(fakeClientSet.AppsClientset.Actions()) != tt.actions) && !tt.wantErr {
-					t.Errorf("expected %v action(s) in GetServiceInstanceLabelValues got %v: %v", tt.actions, len(fakeClientSet.AppsClientset.Actions()), fakeClientSet.AppsClientset.Actions())
-				}
-
-			} else if err == nil && tt.wantErr {
-				t.Error("test failed, expected: false, got true")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, expected: no error, got error: %s", err.Error())
-			}
-
 		})
 	}
 }
@@ -5252,79 +4644,6 @@ func Test_findContainer(t *testing.T) {
 	}
 }
 
-func TestWaitAndGetDC(t *testing.T) {
-	type args struct {
-		name       string
-		annotation string
-		value      string
-		dc         appsv1.DeploymentConfig
-		timeout    time.Duration
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-		actions int
-	}{
-		{
-			name: "Case 1 - Check that the function actually works",
-			args: args{
-				name:       "foo",
-				annotation: "app.kubernetes.io/component-source-type",
-				value:      "git",
-				dc: *fakeDeploymentConfig("foo", "bar",
-					[]corev1.EnvVar{}, []corev1.EnvFromSource{}, t),
-				timeout: 3 * time.Second,
-			},
-			wantErr: false,
-			actions: 1,
-		},
-		{
-			name: "Case 2 - Purposefully timeout / error",
-			args: args{
-				name:       "foo",
-				annotation: "app.kubernetes.io/component-source-type",
-				value:      "foobar",
-				dc: *fakeDeploymentConfig("foo", "bar",
-					[]corev1.EnvVar{}, []corev1.EnvFromSource{}, t),
-				timeout: 3 * time.Second,
-			},
-			wantErr: true,
-			actions: 1,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fakeClient, fakeClientSet := FakeNew()
-			fkWatch := watch.NewFake()
-			go func() {
-				fkWatch.Modify(&tt.args.dc)
-			}()
-			fakeClientSet.AppsClientset.PrependWatchReactor("deploymentconfigs", func(action ktesting.Action) (handled bool, ret watch.Interface, err error) {
-				return true, fkWatch, nil
-			})
-			// Run function WaitAndGetDC
-			_, err := fakeClient.WaitAndGetDC(tt.args.name, 0, tt.args.timeout, func(*appsv1.DeploymentConfig, int64) bool {
-				return !tt.wantErr
-			})
-			// Error checking WaitAndGetDC
-			if !tt.wantErr == (err != nil) {
-				t.Errorf(" client.WaitAndGetDC() unexpected error %v, wantErr %v", err, tt.wantErr)
-			}
-			if err == nil && !tt.wantErr {
-				// Check to see how many actions are being ran
-				if (len(fakeClientSet.AppsClientset.Actions()) != tt.actions) && !tt.wantErr {
-					t.Errorf("expected %v action(s) in WaitAndGetDC got %v: %v", tt.actions, len(fakeClientSet.AppsClientset.Actions()), fakeClientSet.AppsClientset.Actions())
-				}
-			} else if err == nil && tt.wantErr {
-				t.Error("test failed, expected: false, got true")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, expected: no error, got error: %s", err.Error())
-			}
-		})
-	}
-}
-
 func TestCreateBuildConfig(t *testing.T) {
 	type args struct {
 		commonObjectMeta metav1.ObjectMeta
@@ -5572,63 +4891,6 @@ func TestGetMatchingPlans(t *testing.T) {
 			t.Errorf("test failed, expected %v, got %v", expected, plans)
 		}
 	})
-}
-
-func TestStartDeployment(t *testing.T) {
-	tests := []struct {
-		name           string
-		deploymentName string
-		wantErr        bool
-	}{
-		{
-			name:           "Case 1: Testing valid name",
-			deploymentName: "ruby",
-			wantErr:        false,
-		},
-		{
-			name:           "Case 2: Testing invalid name",
-			deploymentName: "",
-			wantErr:        true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			fkclient, fkclientset := FakeNew()
-
-			fkclientset.AppsClientset.PrependReactor("create", "deploymentconfigs", func(action ktesting.Action) (bool, runtime.Object, error) {
-				deploymentConfig := appsv1.DeploymentConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: tt.deploymentName,
-					},
-				}
-				return true, &deploymentConfig, nil
-			})
-
-			_, err := fkclient.StartDeployment(tt.deploymentName)
-			if !tt.wantErr == (err != nil) {
-				t.Errorf(" client.StartDeployment(string) unexpected error %v, wantErr %v", err, tt.wantErr)
-			}
-
-			if err == nil {
-
-				if len(fkclientset.AppsClientset.Actions()) != 1 {
-					t.Errorf("expected 1 action in StartDeployment got: %v", fkclientset.AppsClientset.Actions())
-				} else {
-					startedDeployment := fkclientset.AppsClientset.Actions()[0].(ktesting.CreateAction).GetObject().(*appsv1.DeploymentRequest)
-
-					if startedDeployment.Name != tt.deploymentName {
-						t.Errorf("deployment name is not matching to expected name, expected: %s, got %s", tt.deploymentName, startedDeployment.Name)
-					}
-
-					if startedDeployment.Latest == false {
-						t.Errorf("deployment is not set to latest")
-					}
-				}
-			}
-		})
-	}
 }
 
 func TestGetPortsFromBuilderImage(t *testing.T) {
