--- pkg/occlient/pods.go
@@ -1,22 +1,0 @@
-package occlient
-
-import (
-	"fmt"
-	componentlabels "github.com/openshift/odo/pkg/component/labels"
-	"github.com/openshift/odo/pkg/util"
-	"github.com/pkg/errors"
-	corev1 "k8s.io/api/core/v1"
-)
-
-func (c *Client) GetPodUsingComponentName(componentName, appName string) (*corev1.Pod, error) {
-	componentLabels := componentlabels.GetLabels(componentName, appName, false)
-	componentSelector := util.ConvertLabelsToSelector(componentLabels)
-	dc, err := c.GetOneDeploymentConfigFromSelector(componentSelector)
-	if err != nil {
-		return nil, errors.Wrap(err, "unable to get deployment for component")
-	}
-	// Find Pod for component
-	podSelector := fmt.Sprintf("deploymentconfig=%s", dc.Name)
-
-	return c.GetOnePodFromSelector(podSelector)
-}
